# swagger.yaml
openapi: 3.0.0
info:
  title: FEMCS API
  description: Fire Extinguisher Management and Compliance System
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local API Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password]
              properties:
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                role: { type: string, enum: [admin, technician, inspector, safety_officer] }
      responses:
        '201': { description: User registered }
        '409': { description: Email already registered }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Login successful, returns JWT token }
        '401': { description: Invalid credentials }
...
  /auth/users:
    get:
      tags: [Auth]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200': { description: List of users }
...